Fix PostgreSQL type error - use numeric instead of float for ROUND()
CRITICAL BUG FIXED:
PostgreSQL's ROUND(value, precision) function requires numeric type,
but we were using ::float which creates double precision type.

ERROR:
  function round(double precision, integer) does not exist
  LINE: ROUND((ust.score::float / ust.max_score::float * 100), 2)

FIX:
  Changed ALL occurrences from ::float to ::numeric

AFFECTED QUERIES:
- /api/hr/results - percentage calculation
- /api/hr/results/stats - overall stats
- /api/manager/results - percentage calculation
- /api/manager/results/stats - overall stats
- Diagnostic script queries

CHANGED:
  ROUND((ust.score::float / ust.max_score::float * 100), 2)
TO:
  ROUND((ust.score::numeric / ust.max_score::numeric * 100), 2)

This was preventing ALL HR and Manager result pages from loading!

WHY THIS HAPPENED:
- ::float works in some databases but PostgreSQL is strict about types
- ROUND() with 2 arguments specifically needs numeric type
- Without precision (just ROUND(value)), float would work

TESTING:
Run: python db/diagnose_hr_results.py
Should now work without errors!

+++ db/diagnose_hr_results.py
@@ -48,9 +48,9 @@ async def diagnose_hr_results():
                 await cur.execute("""
                     SELECT
                         COUNT(*) as total_tests,
-                        AVG(score::float / max_score::float * 100) as avg_percentage,
-                        MIN(score::float / max_score::float * 100) as min_percentage,
-                        MAX(score::float / max_score::float * 100) as max_percentage,
+                        AVG(score::numeric / max_score::numeric * 100) as avg_percentage,
+                        MIN(score::numeric / max_score::numeric * 100) as min_percentage,
+                        MAX(score::numeric / max_score::numeric * 100) as max_percentage,
                         AVG(EXTRACT(EPOCH FROM (completed_at - started_at)) / 60) as avg_duration_minutes
                     FROM user_specialization_tests
                     WHERE completed_at IS NOT NULL
@@ -70,8 +70,8 @@ async def diagnose_hr_results():
                 await cur.execute("""
                     SELECT
                         CASE
-                            WHEN (score::float / max_score::float * 100) >= 67 THEN 'Senior'
-                            WHEN (score::float / max_score::float * 100) >= 34 THEN 'Middle'
+                            WHEN (score::numeric / max_score::numeric * 100) >= 67 THEN 'Senior'
+                            WHEN (score::numeric / max_score::numeric * 100) >= 34 THEN 'Middle'
                             ELSE 'Junior'
                         END as level,
                         COUNT(*) as count
@@ -99,7 +99,7 @@ async def diagnose_hr_results():
                         s.name as specialization,
                         ust.score,
                         ust.max_score,
-                        ROUND((ust.score::float / ust.max_score::float * 100), 2) as percentage
+                        ROUND((ust.score::numeric / ust.max_score::numeric * 100), 2) as percentage
                     FROM user_specialization_tests ust
                     JOIN users u ON ust.user_id = u.id
                     JOIN specializations s ON ust.specialization_id = s.id
